requirements.txt
Django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
pandas==2.2.3
openpyxl==3.1.5
reportlab==4.2.5
